@@ -4,11 +4,13 @@ import androidx.recyclerview.widget.RecyclerView
 import android.view.View
 import android.widget.ProgressBar
 import android.widget.TextView
-import kotlinx.android.synthetic.main.simple_list_item.view.*
+import com.butterflymx.butterflymxapiclient.databinding.SimpleListItemBinding
 
-class SimpleListFragmentVH(itemView: View) : RecyclerView.ViewHolder(itemView) {
-    val root: View = itemView.rootView
-    val title: TextView = itemView.tv_title
-    val subTitle: TextView = itemView.tv_sub_title
-    val progressBar: ProgressBar = itemView.progress_bar
+class SimpleListFragmentVH(
+    val binding: SimpleListItemBinding
+) : RecyclerView.ViewHolder(binding.root) {
+    val root: View get() = binding.root
+    val title: TextView get() = binding.tvTitle
+    val subTitle: TextView get() = binding.tvSubTitle
+    val progressBar: ProgressBar get() = binding.progressBar
 }
