@@ -6,22 +6,30 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.navigation.Navigation
 import com.butterflymx.butterflymxapiclient.R
+import com.butterflymx.butterflymxapiclient.databinding.FeaturesBinding
 import com.butterflymx.butterflymxapiclient.utils.mvp.BaseView
-import kotlinx.android.synthetic.main.features.*
 
 class FeaturesFragment : BaseView() {
 
+    private var _binding: FeaturesBinding? = null
+    private val binding get() = _binding!!
+
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        return inflater.inflate(R.layout.features, container, false)
+        _binding = FeaturesBinding.inflate(layoutInflater)
+        return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        ll_open_door.setOnClickListener { swipeToOpenDoor() }
+        binding.llOpenDoor.setOnClickListener { swipeToOpenDoor() }
     }
 
     private fun swipeToOpenDoor() {
         activity?.let { Navigation.findNavController(it, R.id.my_nav_host_fragment).navigate(R.id.select_unit_fragment) }
     }
 
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
 }
